Move backend fact to entities and store it in database
