+++ src/main/java/com/example/demo/models/businesslogic/Matching.java
@@ -1,7 +1,6 @@
 package com.example.demo.models.businesslogic;
 
-import com.example.demo.models.businesslogic.backend.BackendFact;
-import com.example.demo.models.businesslogic.frontend.FrontAnswerElement;
+import com.example.demo.models.entities.BackendFact;
 import com.example.demo.models.entities.*;
 import com.example.demo.utils.HyperText;
