+++ src/main/java/com/example/demo/models/businesslogic/backend/QuestionBack.java
@@ -1,6 +1,7 @@
 package com.example.demo.models.businesslogic.backend;
 
 import com.example.demo.models.entities.AnswerObject;
+import com.example.demo.models.entities.BackendFact;
 import com.example.demo.models.entities.EnumData.QuestionType;
 import com.example.demo.models.businesslogic.Law;
 import com.example.demo.models.entities.Response;
