+++ src/main/java/com/example/demo/models/businesslogic/backend/SWRLBackend.java
@@ -1,6 +1,7 @@
 package com.example.demo.models.businesslogic.backend;
 
 import com.example.demo.models.businesslogic.Law;
+import com.example.demo.models.entities.BackendFact;
 import com.example.demo.models.entities.LawFormulation;
 import com.example.demo.models.entities.Mistake;
 import gnu.trove.map.hash.THashMap;
